Move skit pure-logic helpers into SyntaxKit; decouple Runner from ArgumentParser [skip ci]
Relocate the 14 skit files that import neither Subprocess nor ArgumentParser
into Sources/SyntaxKit/Execution/, drop their now-unnecessary
`#if canImport(Subprocess)` guards, and expose the cross-target ones at
`package` access (matching the DocumentationHarness pattern). skit gains a
SyntaxKit dependency and imports it where it consumes the moved types.
Also corrects OutputCache's stale "@unchecked Sendable" doc comment.

Decouple Runner from the CLI so it can be driven in-process:
- New RunError typed enum (invalidInput / renderFailed / batchFailed / unexpected).
- Runner drops `import ArgumentParser`; `callAsFunction` is now
  `throws(RunError)` and no longer throws ExitCode or calls `exit()`.
- RunInput.resolve throws RunError instead of ArgumentParser's ValidationError.
- Skit+Run maps each RunError case back to ExitCode/ValidationError in a
  `render(using:)` helper, preserving exact CLI behavior (invalid input -> 64,
  render failure -> subprocess code, batch failure -> 1).

Uses enum-pattern catch clauses to sidestep a Swift 6.3.2 SILGen crash on
the always-true `catch let error as RunError` cast under typed throws.

All 398 tests pass; Periphery reports no unused code.
