+++ Package.swift
@@ -144,6 +144,7 @@ let package = Package(
     .executableTarget(
       name: "skit",
       dependencies: [
+        "SyntaxKit",
         "SyntaxParser",
         .product(name: "SwiftSyntax", package: "swift-syntax"),
         .product(name: "SwiftParser", package: "swift-syntax"),
