+++ Sources/SyntaxKit/Execution/SwiftRunOutcome.swift
@@ -1,5 +1,5 @@
 //
-//  ProcessInfo+SyntaxKitCacheRoot.swift
+//  SwiftRunOutcome.swift
 //  SyntaxKit
 //
 //  Created by Leo Dion.
@@ -27,20 +27,12 @@
 //  OTHER DEALINGS IN THE SOFTWARE.
 //
 
-#if canImport(Subprocess)
+package import Foundation
 
-  import Foundation
-
-  extension ProcessInfo {
-    /// Root for all skit caches: `<XDG_CACHE_HOME>/syntaxkit` when that env
-    /// var is set and non-empty, otherwise `defaultRoot` (typically the
-    /// platform's home-relative cache dir).
-    internal func syntaxKitCacheRoot(default defaultRoot: URL) -> URL {
-      if let xdg = environment["XDG_CACHE_HOME"], !xdg.isEmpty {
-        return URL(fileURLWithPath: xdg).appendingPathComponent("syntaxkit")
-      }
-      return defaultRoot
-    }
-  }
-
-#endif
+/// Either the spawned `swift` ran to completion (success or failure) or
+/// the watchdog elapsed first. The completed payload is normalized to the
+/// shape callers want regardless of platform.
+package enum SwiftRunOutcome: Sendable {
+  case completed(exitCode: Int32, stdout: Data, stderr: String)
+  case timedOut
+}
